requirements.txt
arrow==0.7.0
blessed==1.14.1
coverage==4.0.3
decorator==4.0.6
dj-database-url==0.3.0
Django==1.8.8
django-bootstrap-pagination==1.5.1
django-debug-toolbar==1.4
django-extensions==1.6.1
django-picklefield==0.3.2
django-q==0.7.13
django-uuslug==1.1.7
future==0.15.2
gunicorn==19.4.5
ipdb==0.8.1
ipython==4.0.3
ipython-genutils==0.1.0
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5
pudb==2015.4.1
py==1.4.31
Pygments==2.1
pytest==2.8.7
pytest-cov==2.2.1
pytest-django==2.9.1
python-dateutil==2.4.2
python-slugify==1.1.4
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.0
selenium==2.50.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.11
sqlparse==0.1.18
traitlets==4.1.0
tweepy==3.5.0
Unidecode==0.4.18
urwid==1.3.1
wcwidth==0.1.6
wheel==0.24.0
whitenoise==2.0.6
mysqlclient